+++ R/organize_ocupations.R
@@ -3,11 +3,11 @@
 #'Función para clasificar las ocupaciones según las 4 dimensiones del Clasificador Nacional de Ocupaciones
 #' (CNO 2001)
 #'@param base Base individual de uno o más períodos
-#'@details #'disclaimer: El script no es un producto oficial de INDEC.
+#'@details  disclaimer: El script no es un producto oficial de INDEC.
 #'
 #'@examples
 #'
-#'#'
+#'
 #'bases <- dplyr::bind_rows(toybase_individual_2016_03,toybase_individual_2016_04)
 #'bases_clasif <- organize_ocupations(base = bases)
 #'